Bugfix - Column Reordering
Ensures that only columns can be dragged into the table. Dragging a column sets the value of DragSource to it. Therefore, checks if DragSource is null and explicitly sets DragSource to null after dropping. Can still be bypassed by droping a column on an invalid position and afterwards an invalid drop element on the table. But no exception will be thrown anymore, instead the operation is equal to immidiatly dropping the column on the table.